+++ src/include/sll/var_arg.h
@@ -1,0 +1,33 @@
+#ifndef __SLL_VAR_ARG_H__
+#define __SLL_VAR_ARG_H__ 1
+#include <sll/common.h>
+#include <sll/types.h>
+
+
+
+#define SLL_VAR_ARG_LIST_TYPE_C 0
+#define SLL_VAR_ARG_LIST_TYPE_SLL 1
+
+
+
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT const void* sll_var_arg_get(sll_var_arg_list_t* va);
+
+
+
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_integer_t sll_var_arg_get_int(sll_var_arg_list_t* va);
+
+
+
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_float_t sll_var_arg_get_float(sll_var_arg_list_t* va);
+
+
+
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_char_t sll_var_arg_get_char(sll_var_arg_list_t* va);
+
+
+
+__SLL_EXTERNAL void sll_var_arg_get_string(sll_var_arg_list_t* va,sll_string_t* o);
+
+
+
+#endif