+++ src/sll/api/string_api.c
@@ -11,6 +11,7 @@
 #include <sll/string.h>
 #include <sll/types.h>
 #include <sll/util.h>
+#include <sll/var_arg.h>
 #include <stdint.h>
 #include <stdio.h>