+++ src/sll/file.c
@@ -5,6 +5,7 @@
 #include <sll/string.h>
 #include <sll/types.h>
 #include <sll/util.h>
+#include <sll/var_arg.h>
 #include <stdarg.h>